feat: ff-admin is created during app start up
