+++ forge/routes/setup/index.js
@@ -6,16 +6,10 @@
  * @namespace setup
  * @memberof forge.routes
  */
-const crypto = require('crypto')
 const fs = require('fs/promises')
 const path = require('path')
 
 const setupApp = path.join(__dirname, '../../../frontend/dist-setup/setup.html')
-
-const generatePassword = () => {
-    const charList = '0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz~!@-#$'
-    return Array.from(crypto.randomFillSync(new Uint32Array(8))).map(x => charList[x % charList.length]).join('')
-}
 
 module.exports = async function (app) {
     let cachedSetupFile
@@ -24,24 +18,6 @@ module.exports = async function (app) {
         if (app.settings.get('setup:initialised')) {
             reply.redirect('/')
             return
-        }
-
-        if (app.config.create_admin) {
-            if (await app.db.models.User.count() === 0) {
-                const password = process.env.FF_ADMIN_PASSWORD || generatePassword()
-                await app.db.models.User.create({
-                    username: 'ff-admin',
-                    name: 'Default Admin',
-                    email: 'admin@example.com',
-                    email_verified: true,
-                    password,
-                    admin: true,
-                    password_expired: true
-                })
-                app.log.info('[SETUP] Created default Admin User')
-                app.log.info('[SETUP] username: ff-admin')
-                app.log.info(`[SETUP] password: ${password}`)
-            }
         }
 
         const csrfToken = await reply.generateCsrf()
